test: HandlerAdapterRegistryTest 패키지 이동 및 mock 사용
